requirements.txt
streamlit>=1.35.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.0
plotly>=5.22.0
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1
requests>=2.32.0
